+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/web/controller/CertifiedProductControllerTest.java
@@ -210,7 +210,7 @@ public class CertifiedProductControllerTest {
 		InheritedCertificationStatus ics = new InheritedCertificationStatus();
 		ics.setInherits(false);
 		updateRequest.setIcs(ics); // Inherited Status = product.getIcs();
-		updateRequest.setChplProductNumber("15.07.07.2642.EIC04.36.0.1.160402");
+		updateRequest.setChplProductNumber("15.07.07.2642.IC04.36.0.1.160402");
 		try {
 			ListingUpdateRequest listingUpdateRequest = new ListingUpdateRequest();
 			listingUpdateRequest.setListing(updateRequest);
@@ -219,7 +219,7 @@ public class CertifiedProductControllerTest {
 			e.printStackTrace();
 		} catch (ValidationException e) {
 			assertNotNull(e);
-			// ICS is false, 15.07.07.2642.EIC04.36.0.1.160402 shows false ICS. No mismatch = error message
+			// ICS is false, 15.07.07.2642.IC04.36.0.1.160402 shows false ICS. No mismatch = error message
 			assertTrue(e.getErrorMessages().contains("Test Tool 'Transport Test Tool' can not be used for criteria '170.314 (b)(6)', "
 					+ "as it is a retired tool, and this Certified Product does not carry ICS."));
 		}
@@ -233,7 +233,7 @@ public class CertifiedProductControllerTest {
 			e.printStackTrace();
 		} catch (ValidationException e) {
 			assertNotNull(e);
-			// ICS is true, 15.07.07.2642.EIC04.36.0.1.160402 shows false ICS. Mismatch = warning message
+			// ICS is true, 15.07.07.2642.IC04.36.0.1.160402 shows false ICS. Mismatch = warning message
 			assertTrue(e.getErrorMessages().contains("The unique id indicates the product does not have ICS but the value for Inherited Certification Status is true."));
 		}
 		Map<String, Object> classificationType = new HashMap<String, Object>();
@@ -255,7 +255,7 @@ public class CertifiedProductControllerTest {
 			e.printStackTrace();
 		} catch (ValidationException e) {
 			assertNotNull(e);
-			// 2014 certEdition; ICS is false, 15.07.07.2642.EIC04.36.0.1.160402 shows false ICS. No mismatch = error message
+			// 2014 certEdition; ICS is false, 15.07.07.2642.IC04.36.0.1.160402 shows false ICS. No mismatch = error message
 			assertTrue(e.getErrorMessages().contains("Test Tool 'Transport Test Tool' can not be used for criteria '170.314 (b)(6)', "
 					+ "as it is a retired tool, and this Certified Product does not carry ICS."));
 		}
@@ -269,7 +269,7 @@ public class CertifiedProductControllerTest {
 			e.printStackTrace();
 		} catch (ValidationException e) {
 			assertNotNull(e);
-			// 2014 certEdition; ICS is true, 15.07.07.2642.EIC04.36.0.1.160402 shows false ICS. Mismatch = warning message
+			// 2014 certEdition; ICS is true, 15.07.07.2642.IC04.36.0.1.160402 shows false ICS. Mismatch = warning message
 			assertTrue(e.getErrorMessages().contains("The unique id indicates the product does not have ICS but the value for Inherited Certification Status is true."));
 		}
 		
@@ -408,13 +408,13 @@ public class CertifiedProductControllerTest {
 		pcpDTO.setCertificationEdition(certEdition);
 		pcpDTO.setCertificationEditionId(3L); // 1 = 2011; 2 = 2014; 3 = 2015
 		pcpDTO.setIcs(false); // Inherited Status = product.getIcs();
-		pcpDTO.setUniqueId("15.07.07.2642.EIC04.36.0.1.160402");
+		pcpDTO.setUniqueId("15.07.07.2642.IC04.36.0.1.160402");
 		CertifiedProductValidator validator = validatorFactory.getValidator(pcpDTO);
 		if(validator != null) {
 			validator.validate(pcpDTO);
 		}
 		// test 1
-		// ICS is false, 15.07.07.2642.EIC04.36.0.1.160402 shows false ICS. No mismatch = error message
+		// ICS is false, 15.07.07.2642.IC04.36.0.1.160402 shows false ICS. No mismatch = error message
 		assertTrue(pcpDTO.getErrorMessages().contains("Test Tool 'Transport Test Tool' can not be used for criteria '170.314 (b)(6)', "
 					+ "as it is a retired tool, and this Certified Product does not carry ICS."));
 		
@@ -424,7 +424,7 @@ public class CertifiedProductControllerTest {
 		if(validator != null) {
 			validator.validate(pcpDTO);
 		}
-		// ICS is false, 15.07.07.2642.EIC04.36.0.1.160402 shows false ICS. No mismatch = error message
+		// ICS is false, 15.07.07.2642.IC04.36.0.1.160402 shows false ICS. No mismatch = error message
 		assertTrue(pcpDTO.getErrorMessages().contains("Test Tool 'Transport Test Tool' can not be used for criteria '170.314 (b)(6)', "
 					+ "as it is a retired tool, and this Certified Product does not carry ICS."));
 		
@@ -438,7 +438,7 @@ public class CertifiedProductControllerTest {
 		if(validator != null) {
 			validator.validate(pcpDTO);
 		}
-		// ICS is false, 15.07.07.2642.EIC04.36.0.1.160402 shows false ICS. No mismatch = error message
+		// ICS is false, 15.07.07.2642.IC04.36.0.1.160402 shows false ICS. No mismatch = error message
 		assertTrue(pcpDTO.getErrorMessages().contains("Test Tool 'Transport Test Tool' can not be used for criteria '170.314 (b)(6)', "
 					+ "as it is a retired tool, and this Certified Product does not carry ICS."));
 		
@@ -448,7 +448,7 @@ public class CertifiedProductControllerTest {
 		if(validator != null) {
 			validator.validate(pcpDTO);
 		}
-		// ICS is false, 15.07.07.2642.EIC04.36.0.1.160402 shows false ICS. No mismatch = error message
+		// ICS is false, 15.07.07.2642.IC04.36.0.1.160402 shows false ICS. No mismatch = error message
 		assertTrue(pcpDTO.getErrorMessages().contains("Test Tool 'Transport Test Tool' can not be used for criteria '170.314 (b)(6)', "
 					+ "as it is a retired tool, and this Certified Product does not carry ICS."));
 	}
@@ -537,7 +537,7 @@ public class CertifiedProductControllerTest {
 		InheritedCertificationStatus ics = new InheritedCertificationStatus();
 		ics.setInherits(true);
 		updateRequest.setIcs(ics); // Inherited Status = product.getIcs();
-		updateRequest.setChplProductNumber("15.07.07.2642.EIC04.36.0.1.160402");
+		updateRequest.setChplProductNumber("15.07.07.2642.IC04.36.0.1.160402");
 		try {
 			ListingUpdateRequest listingUpdateRequest = new ListingUpdateRequest();
 			listingUpdateRequest.setListing(updateRequest);
@@ -655,7 +655,7 @@ public class CertifiedProductControllerTest {
 		InheritedCertificationStatus ics = new InheritedCertificationStatus();
 		ics.setInherits(true);
 		updateRequest.setIcs(ics); // Inherited Status = product.getIcs();
-		updateRequest.setChplProductNumber("15.07.07.2642.EIC04.36.0.1.160402");
+		updateRequest.setChplProductNumber("15.07.07.2642.IC04.36.0.1.160402");
 		ListingUpdateRequest listingUpdateRequest = new ListingUpdateRequest();
 		listingUpdateRequest.setListing(updateRequest);
 		try{
@@ -759,7 +759,7 @@ public class CertifiedProductControllerTest {
 		pcpDTO.setCertificationEdition(certEdition);
 		pcpDTO.setCertificationEditionId(3L); // 1 = 2011; 2 = 2014; 3 = 2015
 		pcpDTO.setIcs(false); // Inherited Status = product.getIcs();
-		pcpDTO.setUniqueId("15.07.07.2642.EIC04.36.0.1.160402");
+		pcpDTO.setUniqueId("15.07.07.2642.IC04.36.0.1.160402");
 		CertifiedProductValidator validator = validatorFactory.getValidator(pcpDTO);
 		if(validator != null) {
 			validator.validate(pcpDTO);
@@ -857,7 +857,7 @@ public class CertifiedProductControllerTest {
 		pcpDTO.setCertificationEdition(certEdition);
 		pcpDTO.setCertificationEditionId(3L); // 1 = 2011; 2 = 2014; 3 = 2015
 		pcpDTO.setIcs(false); // Inherited Status = product.getIcs();
-		pcpDTO.setUniqueId("15.07.07.2642.EIC04.36.0.1.160402");
+		pcpDTO.setUniqueId("15.07.07.2642.IC04.36.0.1.160402");
 		CertifiedProductValidator validator = validatorFactory.getValidator(pcpDTO);
 		if(validator != null) {
 			validator.validate(pcpDTO);
@@ -955,7 +955,7 @@ public class CertifiedProductControllerTest {
 		pcpDTO.setCertificationEdition(certEdition);
 		pcpDTO.setCertificationEditionId(3L); // 1 = 2011; 2 = 2014; 3 = 2015
 		pcpDTO.setIcs(false); // Inherited Status = product.getIcs();
-		pcpDTO.setUniqueId("15.07.07.2642.EIC04.36.0.1.160402");
+		pcpDTO.setUniqueId("15.07.07.2642.IC04.36.0.1.160402");
 		CertifiedProductValidator validator = validatorFactory.getValidator(pcpDTO);
 		if(validator != null) {
 			validator.validate(pcpDTO);